Fixed Client header styles. Moved fade-out gradient from Services to Clients section.
