+++ src/components/organisms/Clients.jsx
@@ -1,13 +1,34 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
+import { em } from 'polished';
 import { Section, TwoToneHeading } from '../atoms';
 import { ClientList } from '../molecules';
 
+const fadeOut = {
+  value: `${em('150px')}`,
+};
+
 const Wrapper = styled(Section)`
   background: url('img/backgrounds/star-field.png');
-  padding-bottom: 25vw;
+  padding-bottom: ${fadeOut.value};
   position: relative;
+
+  &::after {
+    background: linear-gradient(
+      to bottom,
+      rgba(0, 0, 0, 0) 0%,
+      rgba(0, 0, 0, 1) 100%
+    );
+    bottom: 0;
+    content: '';
+    display: block;
+    height: ${fadeOut.value};
+    left: 0;
+    position: absolute;
+    width: 100%;
+    z-index: 0;
+  }
 `;
 
 const ClientsHeading = styled(TwoToneHeading)`
@@ -34,7 +55,7 @@ class Clients extends React.Component {
     return (
       <Wrapper id="clients">
         <ClientsHeading>
-          We <span>Collaborate</span>
+          <span>We</span> Collaborate
         </ClientsHeading>
         <ClientList>
           {clients.map(client => (
